refactor(#38): member API 요청/응답 타입을 _types → _api/types로 이동
