+++ src/app/member/_api/index.ts
@@ -1,2 +1,3 @@
 export * from "./member.api";
 export * from "./member.queries";
+export * from "./member.types";
