+++ src/app/member/_types/member.types.ts
@@ -1,21 +1,5 @@
-export type Member = {
-  id: number;
-  email: string;
-  isSignUp: boolean;
-  nickname: string;
-  phoneNumber: string;
-  optInMarketing: boolean;
-};
-export type MemberResponse = Member;
-
-export type UpdateMemberRequest = Pick<
-  Member,
-  "nickname" | "phoneNumber" | "optInMarketing"
->;
-export type UpdateMemberResponse = Member;
+import { type Member } from "../_api";
 
 export type NicknameFormValues = Pick<Member, "nickname">;
-export type NicknameCheckRequest = Pick<Member, "nickname">;
 
 export type PhoneNumberFormValues = Pick<Member, "phoneNumber">;
-export type PhoneNumberCheckRequest = Pick<Member, "phoneNumber">;
